refactor(frontend): reorganize JS into pages/, components/, services/, core/
Split flat JS files into proper folder structure following single responsibility:
- core/: api, router, route-protection
- services/: auth.service
- components/: toast, spinner, error-message, layout
- pages/: auth flows, goals, statistics, add-goal, profile, goal-detail

Also fixed sidebar Projects button (text navigates, arrow opens dropdown),
removed dummy data and load real goals from API, added empty-state and
completed-page images, and updated pie chart to use per-goal colors.

Closes #108
