+++ frontend/index.html
@@ -12,7 +12,7 @@
 <body>
     <div id="app"></div>
 
-    <!-- Router Script -->
-    <script type="module" src="js/app-main.js?v=1771446680"></script>
+    <!-- Entry point — reorganized per issue #108 -->
+    <script type="module" src="js/app.js"></script>
 </body>
 </html>
