+++ frontend/js/core/route-protection.js
