+++ frontend/js/pages/dashboard.js
@@ -1,0 +1,10 @@
+/**
+ * Dashboard - redirects to /projects
+ * Extracted from layout.js
+ */
+
+function renderDashboard() {
+    router.navigate('/projects');
+}
+
+window.renderDashboard = renderDashboard;
