+++ frontend/js/pages/goals.js
@@ -1,0 +1,289 @@
+/**
+ * Goals / Projects Pages
+ * From goals-list.js — includes:
+ *   - goal colors (from goal.color or localStorage)
+ *   - sad-folder image for empty state
+ *   - real completed page with happy-calendar image
+ */
+
+// ============================================
+// GOALS API SERVICE
+// ============================================
+
+const goalsService = {
+    async fetchGoals() {
+        return await api.get('/goals/');
+    },
+    async fetchGoal(goalId) {
+        return await api.get(`/goals/${goalId}`);
+    },
+    async createGoal(goalData) {
+        return await api.post('/goals/', goalData);
+    },
+    async updateGoal(goalId, goalData) {
+        return await api.put(`/goals/${goalId}`, goalData);
+    },
+    async deleteGoal(goalId) {
+        await api.delete(`/goals/${goalId}`);
+    }
+};
+
+
+// ============================================
+// HELPERS
+// ============================================
+
+function _getGoalColor(goal) {
+    if (goal.color) return goal.color;
+    const stored = JSON.parse(localStorage.getItem('goalColors') || '{}');
+    return stored[goal.id] || stored[goal.title] || '#00BCD4';
+}
+
+function _fmtMin(durationMin) {
+    const h = Math.floor(durationMin / 60);
+    const m = durationMin % 60;
+    return `${h}h ${String(m).padStart(2, '0')}m`;
+}
+
+function _createLoadingState() {
+    return `
+        <div class="loading-state" style="text-align:center;padding:60px 0;">
+            <div class="spinner-large"></div>
+            <p style="margin-top:16px;color:#999;">Loading...</p>
+        </div>
+    `;
+}
+
+function _createErrorState(message) {
+    return `
+        <div class="error-state" style="text-align:center;padding:60px 0;">
+            <div style="font-size:48px;margin-bottom:16px;">⚠️</div>
+            <h2 style="color:#333;">Failed to load</h2>
+            <p style="color:#999;">${message || 'Something went wrong'}</p>
+            <button onclick="window.location.reload()" class="btn-primary" style="margin-top:16px;">Try Again</button>
+        </div>
+    `;
+}
+
+
+// ============================================
+// GOAL CARD (active goal)
+// ============================================
+
+function createGoalCard(goal) {
+    const color       = _getGoalColor(goal);
+    const progress    = goal.is_complete ? 100 : 0;
+    const durationMin = goal.duration_min || 0;
+    const timeDisplay = durationMin > 0 ? `0h 00m / ${_fmtMin(durationMin)}` : '0h 00m';
+
+    return `
+        <div class="project-list-card" onclick="router.navigate('/goal/${goal.id}')">
+            <button class="project-play-circle" style="background:${color}; box-shadow:0 4px 12px ${color}55;"
+                onclick="event.stopPropagation(); router.navigate('/project/${goal.id}')" title="Start session">
+                <img src="assets/icons/play_vector.svg" alt="Play">
+            </button>
+            <div class="project-list-info">
+                <span class="project-list-name">${goal.title || 'Untitled'}</span>
+                <span class="project-list-time">${timeDisplay}</span>
+                <div class="project-list-bar">
+                    <div class="project-list-bar-fill" style="width:${progress}%; background:${color};"></div>
+                </div>
+            </div>
+            <span class="project-list-percent">${progress}%</span>
+            <img src="assets/icons/next_vector.svg" class="project-list-chevron" alt="">
+        </div>
+    `;
+}
+
+function createGoalsList(goals) {
+    if (!goals || goals.length === 0) {
+        return `
+            <div class="empty-state">
+                <img src="assets/icons/empty-folder.png" alt="No projects" style="width:200px;height:auto;margin-bottom:24px;">
+                <h2>No projects yet</h2>
+                <p>You haven't created any projects yet.<br>Start by adding your first goal to begin tracking your progress.</p>
+                <button onclick="router.navigate('/add-goal')" class="btn-primary" style="margin-top:8px;">Create Goal</button>
+            </div>
+        `;
+    }
+    return `
+        <div class="project-list-outer">
+            ${goals.map(g => createGoalCard(g)).join('')}
+        </div>
+    `;
+}
+
+
+// ============================================
+// COMPLETED GOAL CARD
+// ============================================
+
+function _createCompletedCard(goal) {
+    const color       = _getGoalColor(goal);
+    const durationMin = goal.duration_min || 0;
+    const timeStr     = durationMin > 0 ? `${_fmtMin(durationMin)} / ${_fmtMin(durationMin)}` : '0h 00m / 0h 00m';
+
+    return `
+        <div class="project-list-card" onclick="router.navigate('/goal/${goal.id}')">
+            <div class="project-done-circle" style="background:${color};">
+                <img src="assets/icons/checkmark_icon.svg" alt="Done">
+            </div>
+            <div class="project-list-info">
+                <span class="project-list-name">${goal.title || 'Untitled'}</span>
+                <span class="project-list-time">${timeStr}</span>
+                <div class="project-list-bar">
+                    <div class="project-list-bar-fill" style="width:100%; background:${color};"></div>
+                </div>
+            </div>
+            <span class="project-list-percent" style="background:#E8F5E9;color:#388E3C;">100%</span>
+            <img src="assets/icons/checkmark_icon.svg" class="project-list-chevron" alt="" style="width:20px;height:20px;opacity:0.7;filter:invert(40%) sepia(90%) saturate(400%) hue-rotate(90deg);">
+        </div>
+    `;
+}
+
+
+// ============================================
+// PROJECTS PAGE (active goals)
+// ============================================
+
+async function renderProjectsPageWithGoals() {
+    const appContainer = document.getElementById('app');
+    const content = `
+        <div class="page-header">
+            <h1>Projects</h1>
+        </div>
+        <div class="search-bar-full">
+            <input type="text" class="search-input-pill" placeholder="Search projects...">
+        </div>
+        <div id="projectsContainer">${_createLoadingState()}</div>
+    `;
+    appContainer.innerHTML = createLayout(content, '/projects');
+    attachNavigationListeners();
+
+    try {
+        const goals   = await goalsService.fetchGoals();
+        const active  = goals.filter(g => !g.is_complete && g.status !== 'completed');
+        const el      = document.getElementById('projectsContainer');
+        if (el) el.innerHTML = createGoalsList(active);
+    } catch (err) {
+        const el = document.getElementById('projectsContainer');
+        if (el) el.innerHTML = _createErrorState(err.message);
+        Toast.error('Failed to load projects');
+    }
+}
+
+
+// ============================================
+// COMPLETED PAGE (completed goals)
+// ============================================
+
+let _completedFilter = 'all';
+
+async function renderCompletedPageWithGoals() {
+    const appContainer = document.getElementById('app');
+    const content = `
+        <div class="page-header">
+            <h1>Completed</h1>
+        </div>
+        <div class="search-bar-full">
+            <input type="text" class="search-input-pill" placeholder="Search completed projects...">
+        </div>
+        <div class="completed-filter-row" style="display:flex;align-items:center;gap:8px;margin-bottom:20px;">
+            <div class="insights-tabs">
+                <button class="tab-btn active" onclick="setCompletedFilter('all', this)">All time</button>
+                <button class="tab-btn" onclick="setCompletedFilter('week', this)">Past week</button>
+            </div>
+            <button class="clear-all-btn" onclick="clearCompleted()" style="margin-left:auto;">Clear all</button>
+        </div>
+        <div id="completedContainer">${_createLoadingState()}</div>
+    `;
+    appContainer.innerHTML = createLayout(content, '/completed');
+    attachNavigationListeners();
+    await _loadCompletedGoals();
+}
+
+async function _loadCompletedGoals() {
+    const el = document.getElementById('completedContainer');
+    if (!el) return;
+    try {
+        const goals     = await goalsService.fetchGoals();
+        let completed   = goals.filter(g => g.is_complete || g.status === 'completed');
+
+        const now       = Date.now();
+        const oneWeekMs = 7 * 24 * 60 * 60 * 1000;
+        let filtered    = completed;
+        if (_completedFilter === 'week') {
+            filtered = completed.filter(g => {
+                const ts = g.completed_at || g.updated_at || g.deadline;
+                return ts && (now - new Date(ts).getTime()) <= oneWeekMs;
+            });
+        }
+
+        if (filtered.length === 0) {
+            el.innerHTML = `
+                <div class="empty-state">
+                    <img src="assets/icons/empty-folder.png" alt="Nothing completed" style="width:180px;height:auto;margin-bottom:24px;">
+                    <h2>No completed projects</h2>
+                    <p>Complete a goal to see it here.</p>
+                </div>
+            `;
+            return;
+        }
+
+        const cards     = filtered.map(g => _createCompletedCard(g)).join('');
+        const weekCount = completed.filter(g => {
+            const ts = g.completed_at || g.updated_at || g.deadline;
+            return ts && (now - new Date(ts).getTime()) <= oneWeekMs;
+        }).length;
+
+        el.innerHTML = `
+            <div class="project-list-outer">${cards}</div>
+            ${weekCount > 0 ? `
+            <div class="completed-summary-card" style="
+                background: white; border-radius: 16px; padding: 24px 28px;
+                margin-top: 20px; display: flex; align-items: center;
+                justify-content: space-between; gap: 20px;
+                box-shadow: 0 2px 12px rgba(0,0,0,0.07);
+            ">
+                <div>
+                    <p style="font-size:18px;font-weight:700;color:#333;margin:0 0 8px 0;">
+                        Past week: ${weekCount} project${weekCount !== 1 ? 's' : ''} has been completed!
+                    </p>
+                    <p style="font-size:14px;color:#666;margin:0;">
+                        Great job! You've completed ${weekCount} project${weekCount !== 1 ? 's' : ''} this past week.<br>
+                        Keep up the great work!
+                    </p>
+                </div>
+                <img src="assets/icons/happy-calendar.png" alt="Great job!" style="width:110px;height:auto;flex-shrink:0;">
+            </div>` : ''}
+        `;
+    } catch (err) {
+        if (el) el.innerHTML = _createErrorState(err.message);
+        Toast.error('Failed to load completed projects');
+    }
+}
+
+window.setCompletedFilter = function(filter, btn) {
+    _completedFilter = filter;
+    document.querySelectorAll('.completed-filter-row .tab-btn').forEach(b => b.classList.remove('active'));
+    btn.classList.add('active');
+    _loadCompletedGoals();
+};
+
+window.clearCompleted = function() {
+    Toast.info('Clear all is not yet implemented on the server.');
+};
+
+
+// ============================================
+// EXPORT
+// ============================================
+
+window.goalsService           = goalsService;
+window.createGoalsList        = createGoalsList;
+window.renderProjectsPageWithGoals = renderProjectsPageWithGoals;
+window.renderCompletedPageWithGoals = renderCompletedPageWithGoals;
+
+// Override main entry points
+window.renderProjects  = renderProjectsPageWithGoals;
+window.renderCompleted = renderCompletedPageWithGoals;
